backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
Pillow==9.0.0
django-filter==22.1
reportlab
psycopg2-binary==2.8.6
drf-extra-fields==3.4.1